Support dropping primary key, foreign key, and unique constraints on MySQL via the drop_constraint :type option
MySQL doesn't support the SQL standard DROP CONSTRAINT syntax. It
requires you specify the type of constraint you want to drop. So
previously Sequel's drop_constraint method did not work at all
on MySQL.  With this commit, you can add a :type => :foreign_key,
:primary_key, or :unique option to drop that type of constraint.
Note that there is no way to drop a check constraint, as they don't
exist on MySQL (they are parsed and accepted but ignored).

+++ lib/sequel/adapters/shared/mysql.rb
@@ -154,6 +154,18 @@ module Sequel
           "ALTER TABLE #{quote_schema_table(table)} CHANGE COLUMN #{quote_identifier(op[:name])} #{column_definition_sql(op.merge(opts))}"
         when :drop_index
           "#{drop_index_sql(table, op)} ON #{quote_schema_table(table)}"
+        when :drop_constraint
+          type = case op[:type]
+          when :primary_key
+            return "ALTER TABLE #{quote_schema_table(table)} DROP PRIMARY KEY"
+          when :foreign_key
+            'FOREIGN KEY'
+          when :unique
+            'INDEX'
+          else
+            raise(Error, "must specify constraint type via :type=>(:foreign_key|:primary_key|:unique) when dropping constraints on MySQL")
+          end
+          "ALTER TABLE #{quote_schema_table(table)} DROP #{type} #{quote_identifier(op[:name])}"
         else
           super(table, op)
         end
